% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census}
\alias{census}
\title{Census is a dataset}
\format{
A data frame
\describe{
  \item{SITE}{Indicates number of site. Each site has a unique number from 1 on}
  \item{PLOT}{Number of plot in each site. From 1-12}
  \item{DATE}{The date when the census was conducted}
  \item{RECORD_ID}{Equivalent to sample ID based on code-site-plot-date, to match with other datasets}
  \item{DIAGONAL_PLANT_NUMBER}{number of plants counted following the diagonal transect method in each plot}
  \item{OFF.DIAGONAL_PLANT_NUMBER}{number of plants counted following a central horizontal transect parallel to the long side of the each plot}
  \item{TOTAL_PLANT_NUMBER..OPTIONAL.}{total number of plants counted in the plot}
  \item{MEAN_FRUITS_PER_PLANT..OPTIONAL.}{mean number of fruits per plant}
  \item{SD_FRUITS_PER_PLANT..OPTIONAL.}{Standard deviation of fruit per plant}
  \item{COMMENTS}{Collector comments}
  \item{COMMENTS2}{Collector comments}
  \item{D}{Date in R-readable format}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
census
}
\description{
This dataset consists in the census recorded samples
}
\keyword{datasets}
