% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandates.R
\name{cleandates}
\alias{cleandates}
\title{Clean dates}
\usage{
cleandates(mydate)
}
\arguments{
\item{mydate}{string(8) in the formar yyyymmdd}
}
\value{
date in r format
}
\description{
Takes the format yyyymmdd used in most of the samples id to convert it into an R readable date format
}
