% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createmap.R
\name{ggplot_world_map}
\alias{ggplot_world_map}
\title{Title}
\usage{
ggplot_world_map(
  mapcol = "seashell3",
  backcol = "white",
  countrycol = mapcol,
  xlim = c(-200, +200),
  ylim = c(-60, +80)
)
}
\arguments{
\item{ylim}{}
}
\value{

}
\description{
Title
}
