% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{records}
\alias{records}
\title{Records is a dataset}
\format{
A data frame
\describe{
  \item{CODES}{These codes can take the values "FH"(Flower head)"Albugo"(common pathogen)"SB"(seedbank)"SOIL"(soil sample before experiment)"Ac"
  and indicates the type of sample}
  \item{SITE}{Indicates number of site. Each site has a unique number from 1 on}
  \item{PLOT}{Number of plot in each site. From 1-12. Some instances beyond 12 exist when parallel experiments were conducted}
  \item{DATE}{The date when the sample was coelected}
  \item{SAMPLE_ID}{Unique value given by the code-site-plot-date}
  \item{COMMENTS2}{Comments from the collectors}
  \item{COMMENTS3}{Comments from the collectors}
  \item{D}{Date in R-readable format}
  \item{RECODE}{Code of the field site and sample in a more compatible format with other spreadsheets}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
records
}
\description{
This dataset consists in all the recorded samples
}
\keyword{datasets}
