% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recordssorted}
\alias{recordssorted}
\title{recordssorted is a dataset}
\format{
A data frame
\describe{
  \item{CODES}{These codes can take the values "FH"(Flower head)"Albugo"(common pathogen)"SB"(seedbank)"SOIL"(soil sample before experiment)"Ac"
  and indicates the type of sample}
  \item{SITE}{Indicates number of site. Each site has a unique number from 1 on}
  \item{PLOT}{Number of plot in each site. From 1-12}
  \item{DATE}{The date when the sample was coelected?}
  \item{SAMPLE_ID}{Unique value given by the code-site-plot-date}
  \item{NUMBER_FLOWERS_COLLECTED}{Number of flowers collected in the pot}
  \item{Notes.by.Ru}{}
  \item{D}{Date in date format}
  \item{RECODE}{?}
  \item{id}{Identifier code built from code-site-plot-date columns}
  \item{year}{}
  \item{startyear}{}
  \item{doy}{Day of the year. Example: 01/01 is 1 01/02 is 2}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
recordssorted
}
\description{
This dataset consists in the census recorded samples
}
\keyword{datasets}
