% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_not_reported_or_started}
\alias{sites_not_reported_or_started}
\title{sites_not_reported_or_started is a dataset}
\format{
A data frame
\describe{
  \item{,1}{Responsable name}
  \item{,2}{Site code number}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
sites_not_reported_or_started
}
\description{
This dataset consists in the name and site code number of experiments that needs to be excluded from the analyses because participants never started the experiments
}
\keyword{datasets}
