% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_with_no_success}
\alias{sites_with_no_success}
\title{sites_with_no_success is a dataset}
\format{
A data frame
\describe{
  \item{,1}{Responsable name}
  \item{,2}{Site code number}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
sites_with_no_success
}
\description{
This dataset consists in the name and site code number of experiments that were started but where plants did not establish and reproduce.
}
\keyword{datasets}
