% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitesinfo}
\alias{sitesinfo}
\title{sitesinfo is a dataset}
\format{
A data frame
\describe{
  \item{NAME}{}
  \item{SITE_CODE}{In the other df is just code right? }
  \item{DIARY}{link to a diary where collectors could put comments?}
  \item{EMAIL}{Contact information lead participant}
  \item{EMAIL_2}{Contact information of support people to lead participant}
  \item{COLLABORATOR}{Contact information of participants on site}
  \item{SHIPMENT_ADDRESS}{}
  \item{PACKAGE_SENT.}{Possible answers are: yes, seeds and loggers?}
  \item{PACKAGE_RECEIVED}{Yes or na}
  \item{STARTED_EXPERIMENT}{Date when the experiment was started}
  \item{X2018_SAMPLES}{Whether participants sent samples in 2018 (to be removed in the next version, as this should be in other spreadsheets, or expand to cover other years)}
  \item{SITE_NAME}{Site name given by particpiants}
  \item{LONGITUDE}{Degrees East}
  \item{LATITUDE}{Degrees North}
  \item{ALTITUDE}{Meters above sea level}
  \item{TIME_ZONE}{}
  \item{EXPECTED_SOWING}{Date of expected sowing}
  \item{ACCESS_SOIL}{Whether the participants had access to appropriate soil}
  \item{SOIL_TYPE}{Which soil}
  \item{IBUTON_READER}{Whether participants had access to climate logger (iButton) readers}
  \item{NUMBER_OF_SITES}{Whether participants were running experiments in one or multiple locations.}
  \item{NOTES}{}
  \item{Pathogen.sampling}{Whether participants aimed to join pathogen sampling experiment}
}
}
\source{
\url{https://grenenet.wordpress.com/}
}
\usage{
sitesinfo
}
\description{
This dataset consists in per site contact details of participants, locations of sites, and other logistical information
}
\keyword{datasets}
